+++ src/ExceptionHandlerHelper.php
@@ -41,41 +41,48 @@ class ExceptionHandlerHelper
         $cfg = static::getExceptionHandlerConfig()['map'];
 
         if ($ex instanceof HttpException) {
-            // Check if we have any exception configuration for this particular Http status code.
+            // Check if we have any exception configuration for this particular HTTP status code.
             // This confing entry is guaranted to exist (at least 'default'). Enforced by tests.
             $http_code = $ex->getStatusCode();
             $ex_cfg = $cfg[ HttpException::class ][ $http_code ] ?? null;
             $ex_cfg = $ex_cfg ?? $cfg[ HttpException::class ]['default'];
-            $result = self::processException($ex, $ex_cfg, $http_code);
+            $result = self::processException($ex, /** @scrutinizer ignore-type */ $ex_cfg, $http_code);
         } elseif ($ex instanceof ValidationException) {
             // This entry is guaranted to exist. Enforced by tests.
             $http_code = HttpResponse::HTTP_UNPROCESSABLE_ENTITY;
-            $ex_cfg = $cfg[ HttpException::class ][ $http_code ];
-            $result = self::processException($ex, $ex_cfg, $http_code);
+            $result = self::processException($ex, $cfg[ HttpException::class ][ $http_code ], $http_code);
         }
 
         if ($result === null) {
             // This entry is guaranted to exist. Enforced by tests.
-            $http_code = HttpResponse::HTTP_INTERNAL_SERVER_ERROR;
-            $ex_cfg = $cfg['default'];
-            $result = self::processException($ex, $ex_cfg, $http_code);
+            $result = self::processException($ex, $cfg['default'], HttpResponse::HTTP_INTERNAL_SERVER_ERROR);
         }
 
         return $result;
     }
 
-    protected static function processException(\Exception $ex, array $ex_cfg, int $http_code)
+    /**
+     * Handles given exception and produces valid HTTP response object.
+     *
+     * @param \Exception $ex                 Exception to be handled.
+     * @param array      $ex_cfg             ExceptionHandler's config excerpt related to $ex exception type.
+     * @param int        $fallback_http_code HTTP code to be assigned to produced $ex related response in
+     *                                       case configuration array lacks own `http_code` value.
+     *
+     * @return \Symfony\Component\HttpFoundation\Response
+     */
+    protected static function processException(\Exception $ex, array $ex_cfg, int $fallback_http_code)
     {
         $api_code = $ex_cfg['api_code'];
-        $http_code = $ex_cfg['http_code'] ?? $http_code;
+        $http_code = $ex_cfg['http_code'] ?? $fallback_http_code;
         $msg_key = $ex_cfg['msg_key'] ?? null;
         $msg_enforce = $ex_cfg['msg_enforce'] ?? false;
 
         // No message key, let's get exception message and if there's nothing useful, fallback to built-in one.
-        $error_message = $ex->getMessage();
+        $msg = $ex->getMessage();
         $placeholders = [
             'api_code' => $api_code,
-            'message'  => ($error_message !== '') ? $error_message : '???',
+            'message'  => ($msg !== '') ? $msg : '???',
         ];
 
         // shall we enforce error message?
@@ -83,26 +90,19 @@ class ExceptionHandlerHelper
             // yes, please.
             if ($msg_key === null) {
                 // there's no msg_key configured for this exact code, so let's obtain our default message
-                $error_message = ($msg_key === null) ? static::getErrorMessageForException($ex, $http_code, $placeholders) : Lang::get($msg_key, $placeholders);
+                $msg = ($msg_key === null) ? static::getErrorMessageForException($ex, $http_code, $placeholders)
+                    : Lang::get($msg_key, $placeholders);
             }
         } else {
-            // nothing enforced, handling pipeline: ex_message -> user_defined msg -> http_ex -> default
-            if ($error_message === '') {
-                $error_message = ($msg_key === null) ? static::getErrorMessageForException($ex, $http_code, $placeholders) : Lang::get($msg_key, $placeholders);
+            // nothing enforced, handling pipeline: ex_message -> user_defined_msg -> http_ex -> default
+            if ($msg === '') {
+                $msg = ($msg_key === null) ? static::getErrorMessageForException($ex, $http_code, $placeholders)
+                    : Lang::get($msg_key, $placeholders);
             }
         }
 
         // Lets' try to build the error response with what we have now
-        $result = static::error($ex, $api_code, $http_code, $error_message);
-
-        if ($result === null) {
-            $ex_cfg = $cfg[ HttpException::class ][ $http_code ];
-            $api_code = $ex_cfg['api_code'] ?? BaseApiCodes::EX_VALIDATION_EXCEPTION();
-            $http_code = $ex_cfg['http_code'] ?? $http_code;
-            $result = static::error($ex, $api_code, $http_code, $error_message);
-        }
-
-        return $result;
+        return static::error($ex, $api_code, $http_code, $msg);
     }
 
     /**
@@ -124,7 +124,8 @@ class ExceptionHandlerHelper
         } else {
             // Still got nothing? Fall back to built-in generic message for this type of exception.
             $key = BaseApiCodes::getCodeMessageKey(($ex instanceof HttpException)
-                ? BaseApiCodes::EX_HTTP_EXCEPTION() : BaseApiCodes::NO_ERROR_MESSAGE());
+                ? /** @scrutinizer ignore-deprecated */ BaseApiCodes::EX_HTTP_EXCEPTION()
+                : /** @scrutinizer ignore-deprecated */ BaseApiCodes::NO_ERROR_MESSAGE());
             $error_message = Lang::get($key, $placeholders);
         }
 
@@ -206,29 +207,33 @@ class ExceptionHandlerHelper
             ->build();
     }
 
-    protected
-    static function getExceptionHandlerDefaultConfig(): array
+    /**
+     * Returns default (built-in) exception handler config array.
+     *
+     * @return array
+     */
+    protected static function getExceptionHandlerDefaultConfig(): array
     {
         return [
             'map' => [
                 HttpException::class => [
                     // used by unauthenticated() to obtain api and http code for the exception
                     HttpResponse::HTTP_UNAUTHORIZED         => [
-                        'api_code' => BaseApiCodes::EX_AUTHENTICATION_EXCEPTION(),
+                        'api_code' => /** @scrutinizer ignore-deprecated */ BaseApiCodes::EX_AUTHENTICATION_EXCEPTION(),
                     ],
                     // Required by ValidationException handler
                     HttpResponse::HTTP_UNPROCESSABLE_ENTITY => [
-                        'api_code' => BaseApiCodes::EX_VALIDATION_EXCEPTION(),
+                        'api_code' => /** @scrutinizer ignore-deprecated */ BaseApiCodes::EX_VALIDATION_EXCEPTION(),
                     ],
                     // default handler is mandatory. `default` entry MUST have both `api_code` and `http_code` set.
                     'default'                               => [
-                        'api_code'  => BaseApiCodes::EX_HTTP_EXCEPTION(),
+                        'api_code'  => /** @scrutinizer ignore-deprecated */ BaseApiCodes::EX_HTTP_EXCEPTION(),
                         'http_code' => HttpResponse::HTTP_BAD_REQUEST,
                     ],
                 ],
                 // default handler is mandatory. `default` entry MUST have both `api_code` and `http_code` set.
                 'default'            => [
-                    'api_code'  => BaseApiCodes::EX_UNCAUGHT_EXCEPTION(),
+                    'api_code'  => /** @scrutinizer ignore-deprecated */ BaseApiCodes::EX_UNCAUGHT_EXCEPTION(),
                     'http_code' => HttpResponse::HTTP_INTERNAL_SERVER_ERROR,
                 ],
             ],
@@ -240,8 +245,7 @@ class ExceptionHandlerHelper
      *
      * @return array
      */
-    protected
-    static function getExceptionHandlerConfig(): array
+    protected static function getExceptionHandlerConfig(): array
     {
         return Util::mergeConfig(static::getExceptionHandlerDefaultConfig(),
             \Config::get(ResponseBuilder::CONF_KEY_EXCEPTION_HANDLER, []));